# Specify `default` to speed up symbol lookup